--- pokedex/src/Components/DetalhePokemon.js
@@ -1,14 +1,0 @@
-import React from 'react 
-
-
- function detalhePokemon  (){
-    return(
-
-        <div>
-
-        </div>
-    )
-
-}
-
-export default detalhePokemon
