--- pokedex/src/Pages/Detalhespokemon/DetalhesPokemon.js
@@ -1,11 +1,0 @@
-import react from 'react' 
-
-function DetalhesPokemon (){
-    return(
-        <div>
-            <h1>Fagner </h1>
-        </div>
-    )
-}
-
-export default DetalhesPokemon
