+++ pokedex/src/routes/Router.js
@@ -1,6 +1,5 @@
 import React from "react";
 import ListPokemon from "../Pages/ListPokemon/ListPokemon";
-//import {ListPokemon} from "../Pages/ListPokemon/ListPokemon";
 import {Pokedex} from "../Pages/Pokedex/Pokedex";
 import {DetailsPokemon} from "../Pages/DetailsPokemon/DetailsPokemon";
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
