+++ pokedex/src/routes/coordinator.js
@@ -6,6 +6,6 @@ export const goToDetailsPokemon = (history) => {
     history.push("/")
 }
 
-/*export const goToPokedex = (history) => {
+export const goToPokedex = (history) => {
     history.push("/")
-}*/
+}
